+++ source/TestRenderer/Core/TestPhysicalDevice.cpp
@@ -51,7 +51,7 @@ namespace test_renderer
 
 	ashes::FormatProperties PhysicalDevice::getFormatProperties( ashes::Format fmt )const
 	{
-		return m_formatProperties[size_t( fmt )];
+		return m_formatProperties[fmt];
 	}
 
 	void PhysicalDevice::doInitialise()
